Moved a lot more logic from handler to storage
This makes sense to do on the storage side.
Stuff like opening the file and sniffing for MIME type is more of a
storage thing to do.
Makes it more consistent with the data saving side, too.

Hopefully easier to test as well. We'll see once I start building the
proper tests.